+++ tests/test_default.py
@@ -31,34 +31,6 @@ def test_omero_server_config(Command, Sudo, key, value):
     assert cfg == value
 
 
-def test_inplace_import(Command, File, Sudo):
-    with Sudo('data-importer'):
-        outimport = Command.check_output(
-            '%s %s import -- --transfer=ln_s /data/import/test.fake' %
-            (OMERO, OMERO_LOGIN))
-
-    try:
-        # 5.3 uses a new output format Image:ID
-        imageid = int(outimport.split(':', 1)[1])
-    except IndexError:
-        imageid = int(outimport)
-    assert imageid
-
-    query = ('SELECT concat(ofile.path, ofile.name) '
-             'FROM FilesetEntry AS fse '
-             'JOIN fse.fileset AS fileset '
-             'JOIN fse.originalFile AS ofile '
-             'JOIN fileset.images AS image '
-             'WHERE image.id = %d' % imageid)
-    with Sudo('data-importer'):
-        outhql = Command.check_output(
-            '%s %s hql -q --style plain "%s"' % (OMERO, OMERO_LOGIN, query))
-
-    f = File('/OMERO/ManagedRepository/%s' % outhql.split(',', 1)[1])
-    assert f.is_symlink
-    assert f.linked_to == '/data/import/test.fake'
-
-
 def test_omero_datadir(File):
     d = File('/OMERO')
     assert d.is_directory
